system_data_types.7: srcfix
The contents of each type are a logical block that is indented as
a block.  They are not separate paragraphs that happen to be
indented separately, but a set of continuous paragraphs, all at
the same level, indented as a block from the list entry--the name
of the type--.  Therefore, it makes more sense to use block
indentation, represented by .RS/.RE, rather than indenting each
paragraph separately.  That way it's also easier to further indent
a separate paragraph inside a block, which happens for example in
the case of float_t & double_t.  It's simply much easier now to
use .IP specifically in those cases where you want to indent just
a single paragraph.

Also added an ending separator comment line just after the last
type.

[mtk: minor edits to commit message]

Signed-off-by: Alejandro Colomar <colomar.6.4.3@gmail.com>
Signed-off-by: Michael Kerrisk <mtk.manpages@gmail.com>